Fix pagination not hiding when there's only 1 page

+++ src/Controllers/PostController.php
@@ -101,7 +101,7 @@ class PostController
         return $this->renderer->render($response, 'post-list', [
             'posts' => $parsedPosts,
             'current_page' => $page,
-            'total_pages' => ceil($total / $limit) ?: 1,
+            'total_pages' => (int) ceil($total / $limit) ?: 1,
         ]);
     }
